refactor(coding): split settings page modules
