+++ src/loushang/coding/ui/settings_status_line.py
@@ -1,0 +1,160 @@
+from __future__ import annotations
+
+from collections.abc import Callable
+from dataclasses import dataclass, field
+
+from loushang.coding.ui.settings_common import (
+    SETTINGS_PAGE_THEME,
+    SETTINGS_VALUE_COLUMN,
+    ConfigRow,
+    bool_text,
+    config_items,
+    is_space_event,
+    is_tab_fallback_key,
+    next_bool_value,
+    row_for_key,
+    settings_header,
+)
+from loushang.coding.ui.status_line import (
+    StatusLinePreviewSnapshot,
+    StatusLineSettings,
+    status_line_fields,
+    status_line_separator,
+    status_line_style_mode,
+)
+from loushang.tui import (
+    CursorDeclaration,
+    InputIntent,
+    RenderConstraints,
+    RenderLine,
+    RenderResult,
+    SearchableList,
+    SearchableListSelect,
+    StatusBar,
+)
+
+
+@dataclass(slots=True)
+class StatusLineSettingsPage:
+    statusline_settings: StatusLineSettings
+    statusline_preview: Callable[[], StatusLinePreviewSnapshot]
+    focused: bool = False
+    settings: SearchableList = field(init=False)
+
+    def __post_init__(self) -> None:
+        self.settings = self._make_list(focused=False)
+
+    def focus(self) -> None:
+        self.focused = True
+        self.settings.focus()
+
+    def blur(self) -> None:
+        self.focused = False
+        self.settings.blur()
+
+    def editor_input_target(self) -> object | None:
+        return self.settings.editor_input_target()
+
+    def set_statusline_settings(self, settings: StatusLineSettings, *, preserve_active_key: str = "") -> None:
+        self.statusline_settings = settings
+        self.settings.set_items(config_items(statusline_rows(settings)), preserve_active_key=preserve_active_key)
+
+    def handle_input(self, event: object) -> object:
+        result = self.settings.handle_input(event)
+        if isinstance(result, SearchableListSelect):
+            return self._setting_intent(result.key)
+        if result is not None:
+            return result
+        if self.settings.focus_region == "list" and is_space_event(event):
+            item = self.settings.active_item
+            if item is not None:
+                return self._setting_intent(item.key)
+        if is_tab_fallback_key(event):
+            return True
+        return None
+
+    def render(self, constraints: RenderConstraints) -> RenderResult:
+        if constraints.max_height <= 0:
+            return RenderResult.from_lines([], constraints=constraints)
+        if constraints.max_height <= 6:
+            return self.settings.render(constraints)
+        result = self.settings.render(
+            RenderConstraints(width=constraints.width, max_height=max(1, constraints.max_height - 4))
+        )
+        header = RenderLine(settings_header(constraints.width))
+        preview = _statusline_preview_lines(
+            self.statusline_preview(),
+            self.statusline_settings,
+            width=constraints.width,
+        )
+        rows = [*result.lines[:3], RenderLine(""), header, *result.lines[3:], RenderLine(""), *preview]
+        cursor = result.cursor
+        if cursor is not None and cursor.row >= 3:
+            cursor = CursorDeclaration(row=cursor.row + 2, column=cursor.column)
+        return RenderResult.from_lines(rows[: constraints.max_height], constraints=constraints, cursor=cursor)
+
+    def _make_list(self, *, focused: bool) -> SearchableList:
+        return SearchableList(
+            config_items(statusline_rows(self.statusline_settings)),
+            placeholder="Search status line...",
+            empty_text="No matching status line settings",
+            focused=focused,
+            search_box=True,
+            detail_column=SETTINGS_VALUE_COLUMN,
+            theme=SETTINGS_PAGE_THEME,
+        )
+
+    def _setting_intent(self, key: str) -> InputIntent | None:
+        row = row_for_key(statusline_rows(self.statusline_settings), key)
+        if row is None or row.disabled:
+            return None
+        return InputIntent(kind="setting", text=row.id, note=next_statusline_value(row.id, row.value))
+
+
+def statusline_rows(settings: StatusLineSettings) -> tuple[ConfigRow, ...]:
+    return (
+        ConfigRow("statusline.enabled", "Enabled", bool_text(settings.enabled)),
+        ConfigRow("statusline.field.model", "Model", bool_text(settings.model)),
+        ConfigRow("statusline.field.workspace", "Workspace", bool_text(settings.workspace)),
+        ConfigRow("statusline.field.branch", "Branch", bool_text(settings.branch)),
+        ConfigRow("statusline.field.session", "Session", bool_text(settings.session)),
+        ConfigRow("statusline.field.runtime", "Runtime", bool_text(settings.runtime)),
+        ConfigRow("statusline.field.queue", "Queue", settings.queue),
+        ConfigRow("statusline.field.message", "Message", settings.message),
+        ConfigRow("statusline.separator", "Separator", settings.separator),
+        ConfigRow("statusline.style", "Style", settings.style),
+    )
+
+
+def next_statusline_value(item_id: str, value: str) -> str:
+    if item_id in {
+        "statusline.enabled",
+        "statusline.field.model",
+        "statusline.field.workspace",
+        "statusline.field.branch",
+        "statusline.field.session",
+        "statusline.field.runtime",
+    }:
+        return next_bool_value(value)
+    if item_id in {"statusline.field.queue", "statusline.field.message"}:
+        return {"auto": "true", "true": "false", "false": "auto"}.get(value, value)
+    if item_id == "statusline.separator":
+        return "dot" if value == "pipe" else "pipe"
+    if item_id == "statusline.style":
+        return {"codex-like": "muted", "muted": "plain", "plain": "codex-like"}.get(value, value)
+    return value
+
+
+def _statusline_preview_lines(
+    snapshot: StatusLinePreviewSnapshot,
+    settings: StatusLineSettings,
+    *,
+    width: int,
+) -> tuple[RenderLine, ...]:
+    result = StatusBar(
+        status_line_fields(snapshot, settings),
+        separator=status_line_separator(settings),
+        style_mode=status_line_style_mode(settings),
+    ).render(RenderConstraints(width=width, max_height=1))
+    line = result.lines[0].text if result.lines else ""
+    return (RenderLine("Preview"), RenderLine(line))
